RETAKE Damageable Retake Projectile/AProjectileLauncher

+++ Unity/TowerDefenseWorkshop_Template/Assets/TD/Scripts/Damageable.cs
@@ -5,7 +5,7 @@
 	public class Damageable : MonoBehaviour
 	{
 		[SerializeField]
-		private int _health = 1;
+		private float _health = 1f;
 
 		[SerializeField]
 		private bool _destroyIfKilled = true;
@@ -21,7 +21,7 @@
 		private int _enemyTypeID = 0;
 
 
-		public delegate void DamageableEvent(Damageable caller, int currentHealth, int damageTaken);
+		public delegate void DamageableEvent(Damageable caller, float currentHealth, float damageTaken);
 		private event DamageableEvent _damageTaken = null;
 
 		public event DamageableEvent DamageTaken
@@ -42,7 +42,7 @@
 			return _aimPosition.position;
 		}
 
-		public void TakeDamage(int damage)
+		public void TakeDamage(float damage)
 		{
 			_health -= damage;
